fix copyMixIn (#801)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/801

* fix copyMixIn in trec modules
* fix unittest
* refactor to reuse utils code and correct overridence behavior

Reviewed By: s4ayub

Differential Revision: D41143674

fbshipit-source-id: d95e8d70346a330c1bd8808501b2d92081ea177f
