+++ torchrec/distributed/types.py
@@ -12,6 +12,7 @@ from enum import Enum, unique
 from typing import Any, Callable, Dict, Generic, Iterator, List, Optional, Type, TypeVar
 
 from torch.autograd.profiler import record_function
+from torchrec.types import ModuleNoCopyMixin
 
 try:
     # For python 3.6 and below, GenericMeta will be used by
@@ -498,16 +499,6 @@ class ShardingEnv:
         return cls(world_size, rank, None)
 
 
-class ModuleCopyMixin:
-    """
-    A mixin to allow modules to override copy behaviors in DMP.
-    """
-
-    def copy(self, device: torch.device) -> nn.Module:
-        # pyre-ignore [16]
-        return self.to(device)
-
-
 class FeatureShardingMixIn:
     """
     Feature Sharding Interface to provide sharding-aware feature metadata.
@@ -552,7 +543,7 @@ class ShardedModule(
     abc.ABC,
     nn.Module,
     Generic[CompIn, DistOut, Out, ShrdCtx],
-    ModuleCopyMixin,
+    ModuleNoCopyMixin,
     ModuleShardingMixIn,
 ):
     """
