+++ torchrec/inference/modules.py
@@ -6,7 +6,6 @@
 # LICENSE file in the root directory of this source tree.
 
 import abc
-import copy
 import json
 from dataclasses import asdict, dataclass
 from typing import Any, Dict, List, Optional, Tuple, Type
@@ -16,12 +15,11 @@ import torch.nn as nn
 import torch.quantization as quant
 import torchrec as trec
 import torchrec.quant as trec_quant
-from torchrec.distributed.utils import sharded_model_copy
+from torchrec.distributed.utils import copy_to_device
 from torchrec.modules.embedding_modules import (
     EmbeddingBagCollectionInterface,
     EmbeddingCollectionInterface,
 )
-from torchrec.types import CopyMixIn
 
 
 def quantize_feature(
@@ -69,41 +67,12 @@ def quantize_embeddings(
     )
 
 
-def copy_to_device(
-    module: nn.Module,
-    current_device: torch.device,
-    to_device: torch.device,
-) -> nn.Module:
-
-    with sharded_model_copy(device=None):
-        copy_module = copy.deepcopy(module)
-
-    # Copy only weights with matching device.
-    def _copy_if_device_match(tensor: torch.Tensor) -> torch.Tensor:
-        if tensor.device == current_device:
-            return tensor.to(to_device)
-        return tensor
-
-    # if this is a sharded module, customize the copy
-    if isinstance(copy_module, CopyMixIn):
-        return copy_module.copy(to_device)
-
-    for child_name, child in copy_module.named_children():
-        if not any([isinstance(submodule, CopyMixIn) for submodule in child.modules()]):
-            child_copy = child._apply(_copy_if_device_match)
-        else:
-            child_copy = copy_to_device(child, current_device, to_device)
-        copy_module.register_module(child_name, child_copy)
-    return copy_module
-
-
-class CopyableMixin(CopyMixIn):
+class CopyableMixin(nn.Module):
     def copy(
         self,
         device: torch.device,
     ) -> nn.Module:
         return copy_to_device(
-            # pyre-ignore [6]
             self,
             current_device=torch.device("cpu"),
             to_device=device,
